backend/flask-execution/requirements.txt
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.9.19
awscrt==0.16.7
Babel==2.9.1
blinker==1.6.2
boto3==1.26.165
botocore==1.29.165
certifi==2023.5.7
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.1.0
chevron==0.13.1
click==8.1.3
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
dbus-python==1.2.18
distro==1.5.0
docker==6.1.3
docutils==0.16
ec2-hibinit-agent==1.0.2
Flask==2.3.2
Flask-Cors==4.0.0
gpg==1.15.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
libcomps==0.1.18
lockfile==0.12.2
MarkupSafe==2.1.3
netifaces==0.10.6
oauthlib==3.0.2
packaging==23.1
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
pycparser==2.20
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.1
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==5.4.1
release-notification==1.2
requests==2.31.0
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
s3transfer==0.6.1
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
support-info==1.0
urllib3==1.26.16
wcwidth==0.2.5
websocket-client==1.6.1
Werkzeug==2.3.6
zipp==3.15.0